+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog
 1.8 (unreleased)
 ----------------
 
-- Nothing changed yet.
-
+- Change JS `Faceted` options in the `ready` function so we are sure that
+  Faceted exists.
+  [gbastien]
 
 1.7 (2018-05-25)
 ----------------